+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* **Build 11x** (2014-07-xx)
+  - Added new layout and page method `onInit()` called after components are initialized and before AJAX requests are processed.
+
 * **Build 115** (2014-07-06)
   - Important! All references to *Email* have been changed to *Mail* and renaming may be required in plugins.
   - Console command october:update now supports --core, --plugins and --force options.
@@ -58,7 +61,7 @@
   - Components can now be dragged from the side navigation directly on to the page.
   - Asset maker methods (addJs, addCss, addRss) now use an optional build code, either *core* or a plugin code. This is converted to a version number to ensure updates are not affected by cached assets.
   - Added new method `addComponent()` to Cms Controller. Components can now act as a proxy for other components.
-  - Added new override method `onInit()` to Components, called before AJAX requests are processed.
+  - Added new override method `init()` to Components, called before AJAX requests are processed.
 
 * **Build 90** (2014-05-23)
   - Class `CmsPropertyHelper` has been deprecated, will be removed year > 2014.